feat(cw-substrate-cli): add `rpc` command
