+++ code/xcvm/cosmwasm/cw-substrate-cli/src/args.rs
@@ -1,15 +1,15 @@
-use crate::{new_cmd::NewCommand, substrate_client::SubstrateCommand};
+use crate::{new_cmd::NewCommand, substrate};
 use clap::{Parser, Subcommand};
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 pub struct Args {
-	#[command(subcommand)]
-	pub main_command: Command,
+    #[command(subcommand)]
+    pub main_command: Command,
 }
 
 #[derive(Debug, Subcommand)]
 pub enum Command {
-	New(NewCommand),
-	Substrate(SubstrateCommand),
+    New(NewCommand),
+    Substrate(substrate::Command),
 }
