+++ code/xcvm/cosmwasm/cw-substrate-cli/src/substrate/rpc.rs
@@ -1,0 +1,7 @@
+use clap::Args;
+
+#[derive(Args, Debug)]
+pub struct Command {
+    #[command(subcommand)]
+    pub subcommands: Subcommands,
+}
